Fixed CelExtraChunk wrong datatype in Asprite-File reader. Removed the default padding from the SpriteAtlasBuilder

+++ Editor/Aseprite/Chunks/CelExtraChunk.cs
@@ -5,18 +5,18 @@ namespace Aseprite.Chunks
     public class CelExtraChunk : Chunk
     {
         public uint Flags { get; private set; }
-        public double PreciseX { get; private set; }
-        public double PreciseY { get; private set; }
-        public double Width { get; private set; }
-        public double Height { get; private set; }
+        public float PreciseX { get; private set; }
+        public float PreciseY { get; private set; }
+        public float Width { get; private set; }
+        public float Height { get; private set; }
 
         public CelExtraChunk(uint length, BinaryReader reader) : base(length, ChunkType.CelExtra)
         {
             Flags = reader.ReadUInt32();
-            PreciseX = reader.ReadDouble();
-            PreciseY = reader.ReadDouble();
-            Width = reader.ReadDouble();
-            Height = reader.ReadDouble();
+            PreciseX = reader.ReadSingle();
+            PreciseY = reader.ReadSingle();
+            Width = reader.ReadSingle();
+            Height = reader.ReadSingle();
 
             reader.ReadBytes(16); // For Future
         }
